Problem4 added a bunch of solutions

+++ .g8/form/app/controllers/$model__Camel$Controller.scala
@@ -6,7 +6,7 @@ import play.api.mvc._
 import play.api.data._
 import play.api.data.Forms._
 
-case class $model;format="Camel"$Data(name: String, age: Int)
+case class $model;format="Camel"$Data(name: String, age: BigInt)
 
 // NOTE: Add the following to conf/routes to enable compilation of this class:
 /*
